pyproject.toml
[tool.poetry]
name = "shahmir-notes-generator"
version = "0.1.0"
description = ""
authors = ["syedshahmirsultan <syedshahmirsultan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
youtube-transcript-api = "^0.6.2"
streamlit = "^1.39.0"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
